requirements.txt
annotated-types==0.6.0
anthropic==0.19.1
anyio==4.3.0
asttokens==2.4.1
asyncclick==8.1.7.1
certifi==2024.2.2
charset-normalizer==3.3.2
decorator==5.1.1
distro==1.9.0
dnspython==2.6.1
executing==2.0.1
filelock==3.13.1
fsspec==2024.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.21.4
idna==3.6
ipython==8.22.2
jedi==0.19.1
matplotlib-inline==0.1.6
openai==1.13.3
packaging==23.2
parso==0.8.3
pexpect==4.9.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.6.3
pydantic_core==2.16.3
Pygments==2.17.2
pymongo==4.6.2
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
tokenizers==0.15.2
tqdm==4.66.2
traitlets==5.14.1
typing_extensions==4.10.0
urllib3==2.2.1
wcwidth==0.2.13
